First commit
First commit of the batch add where you can invite a number of users to GitHub using a file containing a list of emails.
This meets with the GitHub rate limit of 5000 transactions per hour, but it hits another rate limit "Over invitation rate limit" which is not defined in any of the GitHub documentation. Further investigation required.